# swagger.yml
openapi: 3.0.2
info:
  title: Task Management
  version: 1.0.0
paths:
  /users:
    summary: Path used to manage the list of users.
    description: >-
      The REST endpoint/path used to list and create zero or more `User` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `User` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              sander:
                value:
                  email: sander123@gmail.com
                  password: pass1234
        required: true
      tags:
        - /users
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                'Invalid email: {reason}':
                  value: 'Invalid email: {reason}'
                Email and password are required:
                  value: Email and password are required
                Password must be at least 8 characters long:
                  value: Password must be at least 8 characters long
                Email must be in a valid format:
                  value: Email must be in a valid format
          description: Client side error
        '409':
          content:
            text/html:
              examples:
                Email already exists:
                  value: Email already exists
          description: The email address provided already exists in the system.
      operationId: createUser
      summary: Create a User
      description: Creates a new instance of a `User`.
  /sessions:
    summary: Path used to manage the list of sessions.
    description: >-
      The REST endpoint/path used to list and create zero or more `Session` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `Session` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
        required: true
      tags:
        - /sessions
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Email and password are required:
                  value: Email and password are required
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Invalid password:
                  value: Invalid password
          description: Unauthorized
        '404':
          content:
            text/html:
              examples:
                User not found:
                  value: User not found
          description: Not Found
      operationId: createSession
      summary: Create a Session
      description: 'Creates a new instance of a `Session`, signing the user in. Returns session ID.'
    delete:
      tags:
        - /sessions
      responses:
        '204':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid authorization header:
                  value: Invalid authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      security:
        -
          Bearer: []
      summary: Destroy a Session
      description: 'Destroys a `Session`, signing the user out. Reads the session ID from Authorization header.'
  /tasks:
    summary: Path used to manage the list of tasks.
    description: >-
      The REST endpoint/path used to list and create zero or more `Task` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    get:
      tags:
        - /tasks
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
          description: Successful response - returns an array of `Task` entities.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid authorization header:
                  value: Invalid authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      security:
        -
          Bearer: []
      operationId: getNotes
      summary: List All Tasks
      description: Gets a list of all `Task` entities.
    post:
      requestBody:
        description: A new `Task` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
            examples:
              Task:
                value:
                  content: some text
                  title: some text
        required: true
      tags:
        - /tasks
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid authorization header:
                  value: Invalid authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      security:
        -
          Bearer: []
      operationId: createNote
      summary: Create a Task
      description: Creates a new instance of a `Task`.
  '/tasks/{id}':
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
            examples:
              PutTask:
                value:
                  content: some text
                  title: some text
        required: true
      tags:
        - /tasks
      parameters:
        -
          examples:
            '1':
              value: '73'
          name: id
          description: The unique identifier of the `Task` to be updated.
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Successful response.
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid authorization header:
                  value: Invalid authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
        '404':
          content:
            text/html:
              examples:
                Task not found:
                  value: Task not found
          description: Task not found
      security:
        -
          Bearer: []
      summary: Update a Task
      description: Updates an existing `Task`.
    delete:
      tags:
        - /tasks
      responses:
        '204':
          description: Successful response.
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid authorization header:
                  value: Invalid authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
        '403':
          content:
            text/html:
              examples:
                Forbidden:
                  value: Forbidden
          description: Forbidden
        '404':
          content:
            text/html:
              examples:
                Task not found:
                  value: Task not found
          description: Not Found
      security:
        -
          Bearer: []
      summary: Delete a Task
      description: Deletes a `Task` entity.
    parameters:
      -
        examples:
          '1':
            value: '1'
        name: id
        description: The unique identifier of the `Task` to be retrieved.
        schema:
          type: integer
        in: path
        required: true
components:
  schemas:
    User:
      title: Root Type for User
      description: ''
      type: object
      properties:
        id:
          format: int32
          description: A field that represents the unique identifier of the user.
          type: integer
          example: '12345'
        email:
          description: A field that represents the email address of the user.
          type: string
          example: john.doe@example.com
        password:
          description: A field that represents the password of the user.
          type: string
          example: MySecurePassword123
      example:
        id: 3
        email: email@example.com
        password: pass123
    Session:
      title: Root Type for Session
      description: ''
      type: object
      properties:
        id:
          type: string
        userId:
          format: int32
          type: integer
      example:
        id: 330d67de-432b-48ce-a617-0bd6774ebcee
        userId: 1
    Credentials:
      title: Root Type for Credentials
      description: ''
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: admin
        password: KollneKollne
    Task:
      title: Root Type for Task
      description: ''
      type: object
      properties:
        content:
          type: string
        id:
          format: int32
          type: integer
        title:
          type: string
        userId:
          format: int32
          type: integer
      example:
        content: This is the content of task 1
        id: 1
        title: Task 1
        userId: 1
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
tags:
  -
    name: /users
    description: ''
  -
    name: /sessions
    description: ''
  -
    name: /tasks
    description: ''